Add OpenUSD Provider based on ImGuiHydraEditor by raph080

+++ src/Lab/ImguiExt.cpp
@@ -725,7 +725,7 @@ bool ImageButton(ImTextureID img,
 
     // Default to using texture ID as ID. User can still push string/integer prefixes.
     // We could hash the size/uv to create a unique ID but that would prevent the user from animating UV.
-    PushID(img);
+    PushID((int) img);
     const ImGuiID id = window->GetID("#image");
     PopID();
